requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.10.6
anthropic==0.49.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytest==8.2.0
httpx==0.27.0